requirements.txt
asyncpg
aiogram
aiohttp
aioredis
python-dotenv
vk_api